refactor(utils): extract shared URL/IO helpers into helpers.py
First step of splitting the ~1700-line utils.py into focused modules. Move the
dependency-free shared helpers -- atomic_write_json, bracket_host,
normalize_cache_key, get_home_directory, truncate_text -- into a new helpers.py
that sits at the bottom of the import graph. utils.py re-exports them (and gains
an __all__) so every `from gopher_mcp.utils import X` keeps working. `os` is kept
importable in utils so the existing `gopher_mcp.utils.os.fsync` patch target is
preserved. No behavior change.
